pyproject.toml
[tool.poetry]
name = "UtopiaForReddit"
version = "1.0.0"
description = "An Accessible Reddit client for Windows and MacOS"
authors = ["Nicklas Tegner <nicklas@accessiware.com>"]

[tool.poetry.dependencies]
python = "^3.6"
appdirs = "^1.4.3"
justupdate = "^1.3.4"
macholib = "^1.14"
altgraph = "^0.17"
pefile = "^2019.4.18"
pyinstaller = "3.6"
pywin32-ctypes = "^0.2.0"
cython = "^0.29.17"
wxasync = "^0.41"
pycryptodomex = "^3.9.7"
requests_cache = "^0.5.2"
praw = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
